Add GHC-8.8 job to travis
Use cabal-install-3.0, as otherwise we won't be able to build
custom setup for GHC-8.8
