+++ cabal-testsuite/Test/Cabal/Monad.hs
@@ -23,8 +23,6 @@ module Test.Cabal.Monad (
     -- * Recording mode
     RecordMode(..),
     testRecordMode,
-    -- * Regex utility (move me somewhere else)
-    resub,
     -- * Derived values from 'TestEnv'
     testCurrentDir,
     testWorkDir,
@@ -58,6 +56,7 @@ module Test.Cabal.Monad (
 
 import Test.Cabal.Script
 import Test.Cabal.Plan
+import Test.Cabal.OutputNormalizer
 
 import Distribution.Simple.Compiler
     ( PackageDBStack, PackageDB(..), compilerFlavor
@@ -67,20 +66,17 @@ import Distribution.Simple.Program.Db
 import Distribution.Simple.Program
 import Distribution.Simple.Configure
     ( configCompilerEx )
-import Distribution.Version
 import Distribution.Text
-import Distribution.Package
 
 import Distribution.Verbosity
 
+import Data.Monoid ((<>), mempty)
 import qualified Control.Exception as E
 import Control.Monad
 import Control.Monad.Trans.Reader
 import Control.Monad.IO.Class
 import Data.Maybe
 import Control.Applicative
-import Data.Monoid
-import qualified Data.Foldable as F
 import System.Directory
 import System.Exit
 import System.FilePath
@@ -89,7 +85,6 @@ import System.IO.Error (isDoesNotExistError)
 import System.IO.Temp (withSystemTempDirectory)
 import System.Process hiding (env)
 import Options.Applicative
-import Text.Regex
 
 data CommonArgs = CommonArgs {
         argCabalInstallPath    :: Maybe FilePath,
@@ -104,7 +99,7 @@ commonArgParser :: Parser CommonArgs
 commonArgParser = CommonArgs
     <$> optional (option str
         ( help "Path to cabal-install executable to test"
-       <> long "with-cabal"
+       Data.Monoid.<> long "with-cabal"
        <> metavar "PATH"
         ))
     <*> optional (option str
@@ -233,7 +228,7 @@ diffProgram = simpleProgram "diff"
 -- | Run a test in the test monad according to program's arguments.
 runTestM :: String -> TestM a -> IO a
 runTestM mode m = withSystemTempDirectory "cabal-testsuite" $ \tmp_dir -> do
-    args <- execParser (info testArgParser mempty)
+    args <- execParser (info testArgParser Data.Monoid.mempty)
     let dist_dir = testArgDistDir args
         (script_dir0, script_filename) = splitFileName (testArgScriptPath args)
         script_base = dropExtensions script_filename
@@ -397,57 +392,6 @@ writeFileNoCR f s =
         hSetNewlineMode h noNewlineTranslation
         hPutStr h s
 
-normalizeOutput :: NormalizerEnv -> String -> String
-normalizeOutput nenv =
-    -- Munge away .exe suffix on filenames (Windows)
-    resub "([A-Za-z0-9.-]+)\\.exe" "\\1"
-    -- Normalize backslashes to forward slashes to normalize
-    -- file paths
-  . map (\c -> if c == '\\' then '/' else c)
-    -- Install path frequently has architecture specific elements, so
-    -- nub it out
-  . resub "Installing (.+) in .+" "Installing \\1 in <PATH>"
-    -- Things that look like libraries
-  . resub "libHS[A-Za-z0-9.-]+\\.(so|dll|a|dynlib)" "<LIBRARY>"
-    -- This is dumb but I don't feel like pulling in another dep for
-    -- string search-replace.  Make sure we do this before backslash
-    -- normalization!
-  . resub (posixRegexEscape (normalizerRoot nenv)) "<ROOT>/"
-  . resub (posixRegexEscape (normalizerTmpDir nenv)) "<TMPDIR>/"
-  . appEndo (F.fold (map (Endo . packageIdRegex) (normalizerKnownPackages nenv)))
-    -- Look for foo-0.1/installed-0d6...
-    -- These installed packages will vary depending on GHC version
-    -- Makes assumption that installed packages don't have numbers
-    -- in package name segment.
-    -- Apply this before packageIdRegex, otherwise this regex doesn't match.
-  . resub "([a-zA-Z]+(-[a-zA-Z])*)-[0-9]+(\\.[0-9]+)*/installed-[A-Za-z0-9.]+"
-          "\\1-<VERSION>/installed-<HASH>..."
-    -- Normalize architecture
-  . resub (posixRegexEscape (display (normalizerPlatform nenv))) "<ARCH>"
-    -- Some GHC versions are chattier than others
-  . resub "^ignoring \\(possibly broken\\) abi-depends field for packages" ""
-    -- Normalize the current GHC version.  Apply this BEFORE packageIdRegex,
-    -- which will pick up the install ghc library (which doesn't have the
-    -- date glob).
-  . (if normalizerGhcVersion nenv /= nullVersion
-        then resub (posixRegexEscape (display (normalizerGhcVersion nenv))
-                        -- Also glob the date, for nightly GHC builds
-                        ++ "(\\.[0-9]+)?")
-                   "<GHCVER>"
-        else id)
-  where
-    packageIdRegex pid =
-        resub (posixRegexEscape (display pid) ++ "(-[A-Za-z0-9.-]+)?")
-              ((display (packageName pid)) ++ "-<VERSION>")
-
-data NormalizerEnv = NormalizerEnv {
-        normalizerRoot :: FilePath,
-        normalizerTmpDir :: FilePath,
-        normalizerGhcVersion :: Version,
-        normalizerKnownPackages :: [PackageId],
-        normalizerPlatform :: Platform
-    }
-
 mkNormalizerEnv :: TestM NormalizerEnv
 mkNormalizerEnv = do
     env <- getTestEnv
@@ -468,16 +412,6 @@ mkNormalizerEnv = do
         normalizerPlatform
             = testPlatform env
     }
-
-posixSpecialChars :: [Char]
-posixSpecialChars = ".^$*+?()[{\\|"
-
-posixRegexEscape :: String -> String
-posixRegexEscape = concatMap (\c -> if c `elem` posixSpecialChars then ['\\', c] else [c])
-
-resub :: String {- search -} -> String {- replace -} -> String {- input -} -> String
-resub search replace s =
-    subRegex (mkRegex search) s replace
 
 requireProgramM :: Program -> TestM ConfiguredProgram
 requireProgramM program = do
