+++ cabal-testsuite/cabal-testsuite.cabal
@@ -26,7 +26,7 @@ common shared
   default-language: Haskell2010
 
   build-depends:
-    , base >= 4.6 && <4.13
+    , base >= 4.6 && <4.14
     -- this needs to match the in-tree lib:Cabal version
     , Cabal == 3.0.0.0
 
@@ -36,14 +36,15 @@ library
   import: shared
 
   exposed-modules:
-    Test.Cabal.Workdir
-    Test.Cabal.Script
-    Test.Cabal.Run
+    Test.Cabal.CheckArMetadata
+    Test.Cabal.Monad
+    Test.Cabal.OutputNormalizer
     Test.Cabal.Plan
     Test.Cabal.Prelude
+    Test.Cabal.Run
+    Test.Cabal.Script
     Test.Cabal.Server
-    Test.Cabal.Monad
-    Test.Cabal.CheckArMetadata
+    Test.Cabal.Workdir
 
   other-modules:
     Test.Cabal.ScriptEnv0
