+++ travis/binaries/.travis.yml
@@ -12,7 +12,7 @@ before_install:
  - export PATH=$HOME/bin:$PATH
  - export PATH=$HOME/.cabal/bin:$PATH
  - export PATH=$HOME/.local/bin:$PATH
- - export PATH=/opt/cabal/2.4/bin:$PATH
+ - export PATH=/opt/cabal/3.0/bin:$PATH
  - export PATH=/opt/happy/1.19.5/bin:$PATH
  - export PATH=/opt/alex/3.1.7/bin:$PATH
  - ./travis-install.sh
